12/7 백엔드 재희 리퀘스트 요청 승인

+++ src/main/java/com/example/travelback/board/controller/Boardcontroller.java
@@ -3,12 +3,17 @@ package com.example.travelback.board.controller;
 
 import com.example.travelback.board.domain.Board;
 import com.example.travelback.board.service.BoardService;
+import com.example.travelback.user.dto.Member;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 @RestController
 @RequiredArgsConstructor
@@ -19,8 +24,9 @@ public class Boardcontroller {
     private  final BoardService service;
 
     @GetMapping("list")
-    public List<Board> list(){
-        return service.list();
+    public Map<String, Object> list(@RequestParam(value = "p" ,defaultValue = "1")Integer page,
+                                    @RequestParam(value = "k" ,defaultValue = "")String keyword){
+        return service.list(page,keyword);
     }
 
 
@@ -28,8 +34,13 @@ public class Boardcontroller {
 
 
     @PostMapping("add")
-    public ResponseEntity add(@RequestBody Board board){
-        System.out.println("Board = " + board);
+    public ResponseEntity add( Board board,
+                               @RequestParam(value = "uploadFiles[]", required = false) MultipartFile[] files,
+                              @SessionAttribute(value = "login" ,required = false)Member login) throws IOException {
+        // 로그인 여부확인
+        if(login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
         // 검증하기
         if (!service.validate(board)){
@@ -37,7 +48,7 @@ public class Boardcontroller {
         }
 
         // 글 추가 하기
-        if (service.add(board)){
+        if (service.add(board,files,login)){
         return     ResponseEntity.ok().build();
         }
         else {
@@ -45,12 +56,6 @@ public class Boardcontroller {
         }
     }
 
-
-    // api/board/list?p=4
-//    @GetMapping("list")
-//    public Map<String ,Object> list(@RequestParam(value = "p",defaultValue = "1")Integer page){
-//        return service.list(page);
-//    }
 
 
     @GetMapping("id/{id}")
@@ -60,7 +65,19 @@ public class Boardcontroller {
 
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id){
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute(value = "login",required = false)Member login){
+
+        //401 권한 없음 에러
+        if (login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        //403 서버에서 거절
+        if (!service.hasAccess(id,login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+
          if (service.remove(id)){
              return ResponseEntity.ok().build();
          }else {
@@ -69,11 +86,21 @@ public class Boardcontroller {
     }
 
     @PutMapping("edit")
-    public  void update(
-            @RequestBody Board board){
-        System.out.println("board = " + board);
+    public  ResponseEntity update(
+            @RequestBody Board board, @SessionAttribute (value = "login",required = false)Member login){
 
-        service.update(board);
+        if (login==null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        if (!service.hasAccess(board.getId(),login)){
+            return  ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        if (service.update(board)) {
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.internalServerError().build();
+        }
 
     }
